[libc.search] [PATCH 3/6] add `search.h` specifications
This patch initiates definitions for search.h as specified in POXIS.1-2001. Because this patch means to
be a part of a larger patch of hashtable implementation, The header definition is not complete,
only hsearch, hdelete, hcreate, ACTION, ENTRY (POSIX standards), and hsearch_r, hdetele_r, hcreate_r,
struct hsearch_data (GNU extensions) are added.